requirements.txt
numpy
plotly
pillow
pyopencl
